Revert "Useless cast"
This reverts commit 7e95963940c6dc5e0cfe46bd59bb9119d1fa19a1.

+++ include/spdlog/details/os-inl.h
@@ -407,7 +407,7 @@ SPDLOG_INLINE int pid() SPDLOG_NOEXCEPT
 #ifdef _WIN32
     return static_cast<int>(::GetCurrentProcessId());
 #else
-    return ::getpid();
+    return static_cast<int>(::getpid());
 #endif
 }
